Update integrity.py
Fixed an issue where it would constantly update the user that changelog.txt's hash is changing, when it should be. As it's supposed to be updated.

+++ integrity.py
@@ -47,43 +47,43 @@ def comp_baseline():
     # Begin (continuously) monitoring files with saved baseline
     if os.path.exists(jsfile): # Load file|hash from baseline.txt and store them in a dictionary
         # Exclude the baseline that is currently being read from
-        # Next Step: Instead of repeatedly saying a file has been changed
-        # Say it only once, add that filename and original hash to a log file
-        # Change the baseline to include the hash of this new file
+        # Next Step:
+        # Don't include Changelog in the scanning. Because it's constantly changing it'll be infinite
         while True:
             for subdir, dirs, files in os.walk(directory):
                 for filename in files:
                     fpath = os.path.join(subdir, filename)
-                    jf = open(jsfile, 'r')
-                    # Declares the json data as a value
-                    data = json.load(jf)
-                    jf.close()
-                    for i in data:
-                        # Checks if the current file matches the filename in the json data
-                        if i == str(fpath):
-                            with open(fpath, 'rb') as tf:
-                                test_hash = hashlib.sha256(tf.read())
-                                # The file hash in the json file is compared to the current hash of the file being scanned
-                                # If the hashes do not match, a log will be updated with the changed log
-                                # and the json file will be updated.
-                                if test_hash.hexdigest() == data[i]:
-                                    pass
-                                # This elif is because this particular hash always appears in the log and baseline before the actual specified file's hash
-                                # Temporary, dirty solution
-                                elif test_hash.hexdigest() != "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855" and test_hash.hexdigest() != data[i]:
-                                    print(data[i], '\n')
-                                    print(test_hash.hexdigest())
-                                    print(f"{i!r} has a non-matching hash.\n")
-                                    with open("changelog.txt", 'a') as lf:
-                                        lf.write(f'{i} : {test_hash.hexdigest()}\nFile was changed at {datetime.datetime.now()}\n\n')
-                                    with open(jsfile, 'r+') as jf:
-                                        data[i] = test_hash.hexdigest()
-                                        jf.seek(0)
-                                        jf.write(json.dumps(data))
-                                        jf.truncate()
-                                tf.close()
-                # Consider scanning on an hourly or daily schedule
-                # instead of a continuous real time scan
+                    if 'File_Integrity\changelog.txt' not in fpath:
+                        jf = open(jsfile, 'r')
+                        # Declares the json data as a value
+                        data = json.load(jf)
+                        jf.close()
+                        for i in data:
+                            # Checks if the current file matches the filename in the json data
+                            if i == str(fpath):
+                                with open(fpath, 'rb') as tf:
+                                    test_hash = hashlib.sha256(tf.read())
+                                    # The file hash in the json file is compared to the current hash of the file being scanned
+                                    # If the hashes do not match, a log will be updated with the changed log
+                                    # and the json file will be updated.
+                                    if test_hash.hexdigest() == data[i]:
+                                        pass
+                                    # This elif is because this particular hash always appears in the log and baseline before the actual specified file's hash
+                                    # Temporary, dirty solution
+                                    elif test_hash.hexdigest() != "e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855" and test_hash.hexdigest() != data[i]:
+                                        print(data[i], '\n')
+                                        print(test_hash.hexdigest())
+                                        print(f"{i!r} has a non-matching hash.\n")
+                                        with open("changelog.txt", 'a') as lf:
+                                            lf.write(f'{i} : {test_hash.hexdigest()}\nFile was changed at {datetime.datetime.now()}\n\n')
+                                        with open(jsfile, 'r+') as jf:
+                                            data[i] = test_hash.hexdigest()
+                                            jf.seek(0)
+                                            jf.write(json.dumps(data))
+                                            jf.truncate()
+                                    tf.close()
+                    # Consider scanning on an hourly or daily schedule
+                    # instead of a continuous real time scan
                     
     else:
         print("Baseline file does not exist in local directory.")
